requirements.txt
requests==2.27.1
Pillow==9.0.1
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3
pytest==7.1.1